requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
google-api-python-client==2.87.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
pandas
